fix(report): stop double-counting flooded volume in node continuity error; 64-bit solver counters
Three fixes found by reviewing the Bellinge 1D+2D SEMI_IMPLICIT status report
against a reproduced run:

1. Node continuity error double-counted flooding.
   The report's per-node error computed out = stat_total_outflow_vol +
   stat_vol_flooded + volume, but stat_total_outflow_vol already includes the
   overflow term for non-ponding flooded nodes (matches legacy
   massbal.c:587-593: NodeOutflow += overflow when newVolume <= fullVolume).
   Adding stat_vol_flooded again counted every flooded node's water twice — at
   2D-coupled nodes the coupling spill is the dominant flood, so those nodes
   reported ~-48% while globally balanced. Legacy computes the node error as
   1 - NodeOutflow/NodeInflow (stats.c:787-788) with no separate flooded term;
   restore that parity.

2. Negative Face-Kernel Evals on wasm32.
   face_passes_/substeps_run_ and the published RunStats/SimulationContext
   counters were `long` — 32-bit under Emscripten. Past 2^31 evaluations the
   report printed a negative eval count (-512788655 in a 48 h Bellinge run).
   Widen the cumulative counters to long long throughout (solvers,
   SimulationContext, report %lld).

3. Reopen the 1D<->2D coupling-conductance gate for SEMI_IMPLICIT.
   computeCouplingConductances (G = -dQ/dh >= 0, pure damping on the node
   continuity denominator) was gated to EXPLICIT "until the SEMI_IMPLICIT
   denominator sign convention is ruled on". The ruling is settled: sumdqdh is
   accumulated POSITIVE with dQ_net/dH = -sumdqdh and the Crank-Nicolson
   denominator is A + 0.5*dt*sumdqdh, so the positive conductance damps both
   modes identically. Applying it under SEMI_IMPLICIT reduces nonconverged
   routing steps on the Bellinge model (~65% -> ~53% at 48 h).

+++ src/engine/2d/solver/ExplicitInertialSolver.hpp
@@ -147,9 +147,9 @@ private:
     /// Lazy-source landing without the O(nt) rebuild (advance boundaries
     /// between rebuild cadences).
     void lazySourcesOnly(double t);
-    long   substeps_run_ = 0;           ///< cumulative substeps (whole run)
-    long   face_passes_  = 0;           ///< cumulative face-kernel passes
-    long   last_steps_   = 0;           ///< substeps in the last advance()
+    long long substeps_run_ = 0;           ///< cumulative substeps (whole run)
+    long long face_passes_  = 0;           ///< cumulative face-kernel passes
+    long      last_steps_   = 0;           ///< substeps in the last advance()
     double last_dt_      = 0.0;
     bool   initialized_  = false;
